requirements.txt
fastapi
uvicorn
pydantic
pandas
fastavro
pyodbc
azure-storage-blob
python-dotenv
os
requests